Added various new functionalities, some still experimental. Commited already as requested by colaborator.
@Maxi: Just run the Method "fillWithPostits" in MySQLUtilities.java to
fill your database with some postits. Remove constrains as needed if it
throws errors.

Added functionality includes under the html
"/userservice/create_user_by_json" a method to unmarshal a json included
in the body of the Rest request. Please refer to "example_user.json" on
how to correctly construct the body.